move croupier as a 1st level package
