+++ src/main/java/jaskell/croupier/ZipScaled.java
@@ -1,4 +1,4 @@
-package jaskell.util.croupier;
+package jaskell.croupier;
 
 import java.util.*;
 import java.util.stream.Collectors;
